python/auth-bot/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-api-python-client==1.6.4
google-cloud-datastore==1.4.0
google-auth==1.2.1
google-auth-oauthlib==0.2.0
google-auth-httplib2==0.0.3
pycrypto==2.6.1
requests==2.18.4
